LPS-61860 move classes to avoid exporting internal package
