+++ modules/portal/portal-search/src/main/java/com/liferay/portal/search/buffer/IndexerRequestBuffer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.internal;
+package com.liferay.portal.search.buffer;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
