+++ modules/portal/portal-search/src/main/java/com/liferay/portal/search/buffer/IndexerRequestBufferOverflowHandler.java
@@ -14,8 +14,6 @@
 
 package com.liferay.portal.search.buffer;
 
-import com.liferay.portal.search.internal.IndexerRequestBuffer;
-
 /**
  * @author Michael C. Han
  */
