+++ modules/portal/portal-search/src/main/java/com/liferay/portal/search/internal/buffer/BufferedIndexerInvocationHandler.java
@@ -20,10 +20,10 @@ import com.liferay.portal.kernel.util.MethodKey;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.model.BaseModel;
 import com.liferay.portal.model.ClassedModel;
+import com.liferay.portal.search.buffer.IndexerRequest;
+import com.liferay.portal.search.buffer.IndexerRequestBuffer;
 import com.liferay.portal.search.buffer.IndexerRequestBufferOverflowHandler;
 import com.liferay.portal.search.configuration.IndexerRegistryConfiguration;
-import com.liferay.portal.search.internal.IndexerRequest;
-import com.liferay.portal.search.internal.IndexerRequestBuffer;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.InvocationHandler;
