+++ modules/portal/portal-search/src/main/java/com/liferay/portal/search/internal/buffer/DefaultIndexerRequestBufferOverflowHandler.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.search.internal.buffer;
 
+import com.liferay.portal.search.buffer.IndexerRequestBuffer;
 import com.liferay.portal.search.buffer.IndexerRequestBufferOverflowHandler;
-import com.liferay.portal.search.internal.IndexerRequestBuffer;
 
 /**
  * @author Michael C. Han
